feat: improve home view
